requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1
psycopg2-binary==2.9.9
